Design changes and memory optimized

+++ Game_Play.cpp
@@ -1,14 +1,14 @@
 #include "Game_Play.h"
 HANDLE Console1 = GetStdHandle(STD_OUTPUT_HANDLE);
 Game_Play::Game_Play() : Position_Cordinates({ 0,0 }), Collsion(0), GamePLay_Difficulty(90000) {}
-void Game_Play::set_Position_COORDS(int x, int y) { Position_Cordinates.X = x;		Position_Cordinates.Y = y; }
+void Game_Play::set_Position_COORDS(short x, short y) { Position_Cordinates.X = x;		Position_Cordinates.Y = y; }
 bool Game_Play::get_Collision_Status() const { return Collsion; }
 COORD Game_Play::get_Player_Position() const { return Position_Cordinates; }
-int Game_Play::getPosition_X() const { return Position_Cordinates.X; }
-int Game_Play::getPosition_Y() const { return Position_Cordinates.Y; }
-void Game_Play::setPosition_Y(int Y) { Position_Cordinates.Y = Y; }
-void Game_Play::Update_Position_X(int x) { Position_Cordinates.X = x; }
-void Game_Play::Update_Position_Y(int y) { Position_Cordinates.Y = y; }
+short Game_Play::getPosition_X() const { return Position_Cordinates.X; }
+short Game_Play::getPosition_Y() const { return Position_Cordinates.Y; }
+void Game_Play::setPosition_Y(short Y) { Position_Cordinates.Y = Y; }
+void Game_Play::Update_Position_X(short x) { Position_Cordinates.X = x; }
+void Game_Play::Update_Position_Y(short y) { Position_Cordinates.Y = y; }
 void Game_Play::Erase_Car()
 {
 	COORD X_and_Y;
@@ -16,7 +16,7 @@ void Game_Play::Erase_Car()
 	X_and_Y.X = Position_Cordinates.X;
 	X_and_Y.Y = Position_Cordinates.Y + 1;
 	SetConsoleCursorPosition(Console1, X_and_Y);
-	for (int loop_iterator = 0; loop_iterator < 25; loop_iterator++)
+	for (short loop_iterator = 0; loop_iterator < 25; loop_iterator++)
 	{
 		if (loop_iterator % 5 == 0)
 		{
@@ -26,6 +26,6 @@ void Game_Play::Erase_Car()
 		cout << " ";
 	}
 }
-int Game_Play::get_Car_width() const { return Game_Car.get_Car_Width(); }
-void Game_Play::set_Difficulty(int Difficulty) { GamePLay_Difficulty = Difficulty; }
-int Game_Play::get_Difficulty() const { return GamePLay_Difficulty; }
+short Game_Play::get_Car_width() const { return Game_Car.get_Car_Width(); }
+void Game_Play::set_Difficulty(short Difficulty) { GamePLay_Difficulty = Difficulty; }
+short Game_Play::get_Difficulty() const { return GamePLay_Difficulty; }
